requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
